move s2s code from modular_models back to modular_s2s_models
